bug fix: upgraded type from float and int to double

+++ tabels-core/src/main/scala/es/ctic/tabels/NumericExpressions.scala
@@ -12,12 +12,12 @@ object NumericFunctions extends FunctionCollection {
     // XPath 2.0 operators on numeric values (6.2)
     // ==========================================    
     
-    val numericAdd = "numeric-add" isDefinedBy { (x : Int, y : Int) => x + y }
-    val numericSubstract = "numeric-substract" isDefinedBy { (x : Int, y : Int) => x - y }
-    val numericMultiply = "numeric-multiply" isDefinedBy { (x : Int, y : Int) => x * y }
-    val numericDivide = "numeric-divide" isDefinedBy { (x : Int, y : Int) => x / y }
-    val numericIntegerDivide = "numeric-integer-divide" isDefinedBy { (x : Int, y : Int) => (x / y).round }
-    val numericMod = "numeric-mod" isDefinedBy { (x : Int, y : Int) => x % y }
+    val numericAdd = "numeric-add" isDefinedBy { (x : Double, y : Double) => x + y }
+    val numericSubstract = "numeric-substract" isDefinedBy { (x : Double, y : Double) => x - y }
+    val numericMultiply = "numeric-multiply" isDefinedBy { (x : Double, y : Double) => x * y }
+    val numericDivide = "numeric-divide" isDefinedBy { (x : Double, y : Double) => x / y }
+    val numericIntegerDivide = "numeric-integer-divide" isDefinedBy { (x : Double, y : Double) => (x / y).round }
+    val numericMod = "numeric-mod" isDefinedBy { (x : Double, y : Double) => x % y }
     // FIXME: numeric-unary-plus is missing
     // FIXME: numeric-unary-minus is missing
     
@@ -25,18 +25,18 @@ object NumericFunctions extends FunctionCollection {
     // XPath 2.0 comparison operators on numeric values (6.3)
     // ==========================================    
     
-    val equal = "numeric-equal" isDefinedBy { (x : Int, y : Int) => x == y }
-    val greaterThan = "numeric-greater-than" isDefinedBy { (x : Int, y : Int) => x > y }
-    val lessThan = "numeric-less-than" isDefinedBy { (x : Int, y : Int) => x < y }
+    val equal = "numeric-equal" isDefinedBy { (x : Double, y : Double) => x == y }
+    val greaterThan = "numeric-greater-than" isDefinedBy { (x : Double, y : Double) => x > y }
+    val lessThan = "numeric-less-than" isDefinedBy { (x : Double, y : Double) => x < y }
 
     // ==========================================    
     // XPath 2.0 functions on numeric values (6.4)
     // ==========================================
 
-    val abs = "abs" isDefinedBy { (x : Float) => scala.math.abs(x)  }
-    val ceiling = "ceiling" isDefinedBy { (x : Float) => scala.math.ceil(x)  }
-    val floor = "floor" isDefinedBy { (x : Float) => scala.math.floor(x)  }
-    val round = "round" isDefinedBy { (x : Float) => scala.math.round(x)  }
+    val abs = "abs" isDefinedBy { (x : Double) => scala.math.abs(x)  }
+    val ceiling = "ceiling" isDefinedBy { (x : Double) => scala.math.ceil(x)  }
+    val floor = "floor" isDefinedBy { (x : Double) => scala.math.floor(x)  }
+    val round = "round" isDefinedBy { (x : Double) => scala.math.round(x)  }
     // FIXME: round-half-to-even is missing
     val int = "int" isDefinedBy { (x : Int) => x  }
     val float = "float" isDefinedBy { (x : Float) => x  }
